+++ game/piece.py
@@ -7,3 +7,7 @@ class Piece:
     
     def __str__(self):
         return f"{self.__class__.__name__}({self.__color__})"
+    
+    def is_opponent_piece(self, other_piece):
+        """Returns True if the other piece is an opponent piece, otherwise False."""
+        return self.__color__ != other_piece.get_color()